requirements.txt
dspy-ai>=2.3.5
requests>=2.28.1
dotenv>=0.0.5
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
langchain>=0.0.267
